[BUG] Fix transactions not always processing
We had a race condition in the processing of a transaction where if the
transaction had not yet been saved before the job ran, it would raise a
SerializationError, and discard.  This happened for even new jobs,
which causes issues.  I'm changing it to always process, and retry,
but to check for the presence of a transaction id first.

I also made a change to only serialize the id of the model, not the
full model itself.