[BUG] Don't run jobs before they're created
This commit continues the work on not trying to run the transaction/
subscription job before it is fully created.  Before, I believed it
was a race condition, but after more digging I see that the call to
enqueue the job was done after_create, which means that it could be
picked up before it has been fully committed to the db.  This change
runs the enqueue after commmit.  Note that means that the callback
will also be fired after an update of the transaction (if that's a
possibility), but guards are in place to not enqueue the same job on
update.