[BUG] Update profile_type, and fix expiration
* The profile type is listed as a string in the db, but is defined as
a enum in the model, which means it is trying to save as an integer.
The result is that all profile_types in the db are nil, and the payment
model is invalid.

* There is a db attribute on payment that is called expiration, as well
as a private method called expiration.  So, the db column can never be
assigned, or accessed without going through the attributes, which is
hacky.