[BUG] Expiration date fix
We were still having an issue with the expiration date showing as a
negative year upon syncing with Authorize.net.  The issue is with the
format_data callback that happens before_validation on the payment
model.  Since the year and month are only attributes of the class, not
fields of the model, they would never come back from Authorize. The
format_data uses those to build the expiration_date, which works
initially from data coming from the user form, but obviously not if
they aren't available.

Instead, I pull the expiration date data from the authorize call, and
parse it into the model.