[BUG] Allow new transactions to sync.
If a transaction (such as a refund) originates within authorize instead of redline, we don't have a record of it at first. The hook system will build a new Transaction, but that transaction sync will fail, since the new Transaction record doesn't have an accountable to which to delegate the profile. If we allow the accountable to be nil, the transactions can sync on transaction_id alone, which will update the transaction with all the information it needs, including accountable_id.